reverted changes made by arjun

+++ api_server/api_server.py
@@ -28,15 +28,6 @@ db = client.certify
 api_auth_key = os.environ.get("API_AUTH_KEY")
 
 
-def authenticate(func):
-    def new_func():
-        provided_key = request.headers.get("API-Auth-Key")
-        if provided_key != api_auth_key:
-            response = {"error": "Invalid API key"}
-            return make_response(json.dumps(response), 400)
-        return func()
-    return new_func
-
 ## STATUS
 
 # Public
@@ -62,8 +53,11 @@ def get_event_info():
 
 # load list of events
 @app.route('/event/list', methods=['GET'])
-@authenticate
 def get_events_list():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     query_result = list(db.events.find({},{"_id":1,"name":1,"issueDt":1}))
     r = make_response(json.dumps(query_result, cls=CustomJSONEncoder))
@@ -72,8 +66,11 @@ def get_events_list():
 
 # Add Event
 @app.route('/event/add', methods=['POST'])
-@authenticate
 def add_event():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     response_body = request.get_json()
 
@@ -91,8 +88,11 @@ def add_event():
 
 # Finalize Event
 @app.route('/event/finalize', methods=['POST'])
-@authenticate
 def finalize_event():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     response_body = request.get_json()
 
@@ -107,8 +107,11 @@ def finalize_event():
 
 # Update Event
 @app.route('/event/update', methods=['POST'])
-@authenticate
 def update_event():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     response_body = request.get_json()
 
@@ -127,8 +130,11 @@ def update_event():
 
 # Delete Event
 @app.route('/event/delete', methods=['DELETE'])
-@authenticate
 def delete_event():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     event_id = ObjectId(request.args.get('event_id'))
     db.events.delete_one({"_id" : event_id})
@@ -155,8 +161,11 @@ def get_participant_info():
 
 # load list of participants
 @app.route('/participant/list', methods=['GET'])
-@authenticate
 def get_participants_list():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     event_id = ObjectId(request.args.get('event_id'))
     query_result = list(db.participants.find({"event_id" : event_id}))
@@ -166,8 +175,11 @@ def get_participants_list():
 
 # add participants
 @app.route('/participant/add', methods=['POST'])
-@authenticate
 def add_participants():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     items = request.get_json()["items"]
     # Convert Participant Id from string to ObjectId
@@ -183,8 +195,16 @@ def add_participants():
 
 # Update Participant
 @app.route('/participant/update', methods=['POST'])
-@authenticate
 def update_participant():
+    # Get the provided API key from the request headers
+    provided_key = request.headers.get("API-Auth-Key")
+
+    # Check if the provided key matches the expected API key
+    if provided_key != api_auth_key:
+        # If the keys don't match, return an error response with status code 401
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
+
     # Get the participant ID, event ID, field, and value from the request arguments
 
     response_body = request.get_json()
@@ -208,8 +228,11 @@ def update_participant():
 
 # Delete Participant
 @app.route('/participant/delete', methods=['DELETE'])
-@authenticate
 def delete_participant():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     
     participant_id = ObjectId(request.args.get('participant_id'))
@@ -223,8 +246,11 @@ def delete_participant():
 
 # Delete Participants in bulk
 @app.route('/participant/delete-batch', methods=['DELETE'])
-@authenticate
 def delete_participants():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     event_id = ObjectId(request.args.get('event_id'))
     db.participants.delete_many({"event_id" : event_id})
@@ -238,8 +264,11 @@ def delete_participants():
 ## PLUGIN
 
 @app.route("/plugin/gen-info", methods=['GET'])
-@authenticate
 def get_gen_info():
+    provided_key = request.headers.get("API-Auth-Key")
+    if provided_key != api_auth_key:
+        response = {"error": "Invalid API key"}
+        return make_response(json.dumps(response), 401)
     
     event_id = ObjectId(request.args.get('event_id'))
